refactor(re-exports)!: re-exported extension and extract mods
