+++ examples/map_err.rs
@@ -20,8 +20,8 @@ use composable_tower_http::{
         },
         header::basic_auth::impls::default_basic_auth_extractor::DefaultBaiscAuthExtractor,
     },
-    extension::layer::ExtensionLayerExt,
-    extract::extractor::ExtractorExt,
+    extension::ExtensionLayerExt,
+    extract::ExtractorExt,
 };
 use http::StatusCode;
