+++ src/extension/mod.rs
@@ -1,2 +1,5 @@
-pub mod layer;
-pub mod service;
+mod layer;
+mod service;
+
+pub use layer::{ExtensionLayer, ExtensionLayerExt};
+pub use service::ExtensionService;
