+++ src/extract/mod.rs
@@ -1,6 +1,25 @@
-pub mod and;
-pub mod any;
-pub mod extracted;
-pub mod extractor;
-pub mod or;
-pub mod sealed_extracted;
+mod and;
+mod any;
+mod chain;
+mod convert;
+mod extracted;
+mod extractor;
+mod map;
+mod optional;
+mod or;
+mod sealed_extracted;
+
+pub use and::{And, AndError};
+pub use any::{Any, AnyError};
+pub use chain::{
+    chain_extractor::{ChainError, ChainExtractor},
+    chainer::Chainer,
+    lite::{AsyncChainLite, ChainLite},
+};
+pub use convert::{AsyncConvert, Convert};
+pub use extracted::Extracted;
+pub use extractor::{Extractor, ExtractorExt};
+pub use map::{AsyncMap, Map, MapError};
+pub use optional::Optional;
+pub use or::{Or, OrError};
+pub use sealed_extracted::SealedExtracted;
