+++ src/lib.rs
@@ -4,7 +4,6 @@
 //! Highly costumizable http utilities built on top of [tower](https://docs.rs/tower/latest/tower/).
 
 pub mod authorize;
-pub mod chain;
 pub mod error;
 pub mod extension;
 pub mod extract;
